Removed redundant stuff
